backend/requirements.txt
fastapi==0.111.1
websockets
numpy==2.0.0
pymongo==4.8.0
pyjwt==2.9.0
passlib[bcrypt]